refactor(devframe)!: move remaining low-level node exports to devframe/internal
Reduce the public `devframe/node` barrel to the server-assembly surface only
(`createHostContext`, `createH3DevframeHost`, `startHttpAndWs`, `createStorage`,
`RpcFunctionsHost`). Relocate the low-level primitives that only first-party
integrations use to the unstable `devframe/internal` surface:

- the instance registry (`registerDevframeInstance` / `listLiveDevframeInstances`
  + `DevframeInstanceRecord` / `DevframeInstanceRegistration`)
- `isObject`, `normalizeHttpServerUrl`

`startHttpAndWs` stays on `devframe/node` — it's the primary programmatic
server entry (used as real code by several plugins, examples, and the hub), not
a low-level util. The inspect plugin and the hub host examples now import the
instance registry from `devframe/internal`; `@vitejs/devtools` picks up
`isObject`/`normalizeHttpServerUrl` from there too.
