+++ examples/a11y-messages-playground/src/a11y-messages-playground.ts
@@ -1,13 +1,14 @@
 import type { DevframeHubContext } from '@devframes/hub/node'
 import type { ClientScriptEntry } from '@devframes/hub/types'
 import type { DevframeDefinition, DevframeHost } from 'devframe'
-import type { DevframeInstanceRegistration } from 'devframe/node'
+import type { DevframeInstanceRegistration } from 'devframe/internal'
 import type { Plugin, ResolvedConfig, ViteDevServer } from 'vite'
 import { homedir } from 'node:os'
 import process from 'node:process'
 import { createHubContext, mountDevframe } from '@devframes/hub/node'
 import { DEVFRAME_CONNECTION_META_FILENAME } from 'devframe/constants'
-import { registerDevframeInstance, startHttpAndWs } from 'devframe/node'
+import { registerDevframeInstance } from 'devframe/internal'
+import { startHttpAndWs } from 'devframe/node'
 import { serveStaticNodeMiddleware } from 'devframe/utils/serve-static'
 import { getPort } from 'get-port-please'
 import { join } from 'pathe'
