+++ examples/hub-next/src/client/devframe/next-devframe-hub.ts
@@ -1,14 +1,14 @@
 import type { HubDevframeEntry, HubInstance } from '@devframes/hub/initiate'
 import type { DevframeHubContext } from '@devframes/hub/node'
 import type { DevframeDefinition } from 'devframe'
-import type { DevframeInstanceRegistration } from 'devframe/node'
+import type { DevframeInstanceRegistration } from 'devframe/internal'
 import { homedir } from 'node:os'
 import process from 'node:process'
 import { fileURLToPath } from 'node:url'
 import { defineHubRpcFunction } from '@devframes/hub'
 import { DEVFRAMES_HUB_BASE, initHub } from '@devframes/hub/initiate'
 import { toJsonRenderDockEntry } from '@devframes/json-render/hub'
-import { registerDevframeInstance } from 'devframe/node'
+import { registerDevframeInstance } from 'devframe/internal'
 import { createDashboardView } from 'json-render/dashboard'
 import { dirname, join } from 'pathe'
 import demoDevframe from './demo-devframe'
