+++ examples/hub-vite/src/vite-devframe-hub.ts
@@ -2,14 +2,14 @@ import type { HubDevframeEntry, HubInstance } from '@devframes/hub/initiate'
 import type { DevframeHubContext } from '@devframes/hub/node'
 import type { ClientScriptEntry } from '@devframes/hub/types'
 import type { DevframeDefinition } from 'devframe'
-import type { DevframeInstanceRegistration } from 'devframe/node'
+import type { DevframeInstanceRegistration } from 'devframe/internal'
 import type { Plugin, ResolvedConfig, ViteDevServer } from 'vite'
 import { Server as NodeHttpServer } from 'node:http'
 import { homedir } from 'node:os'
 import process from 'node:process'
 import { defineHubRpcFunction } from '@devframes/hub'
 import { DEVFRAMES_HUB_BASE, initHub } from '@devframes/hub/initiate'
-import { registerDevframeInstance } from 'devframe/node'
+import { registerDevframeInstance } from 'devframe/internal'
 import { join } from 'pathe'
 
 export interface ViteDevframeHubOptions {
