+++ plans/032-0.9-public-api-surface-reduction.md
@@ -95,12 +95,18 @@ Net effect (after maintainer follow-ups):
 - Dead-subpath removals are `utils/promise` and `utils/scope` only. `ws-bun` is
   kept (hub uses it); `utils/hash` is kept (maintainer request); `embedded` is
   kept (maintainer request — a named, discoverable adapter).
-- The `devframe/node` trim drops **12** exports again — but the three
-  hub-only low-level APIs (`createContextRpcServer`, `DevframeAgentHost`,
-  `coerceAgentPositionalArgs`) are **relocated** to a new **`devframe/internal`**
-  entry (an explicitly-unstable cross-package surface) rather than kept on the
-  public `devframe/node`, and the hub imports them from there. The other 9 stay
-  fully internal (relative imports only).
+- The public `devframe/node` barrel is reduced to the server-assembly surface
+  only: `createHostContext`, `createH3DevframeHost`, `startHttpAndWs`
+  (+ `StartedServer`), `createStorage`, and the `RpcFunctionsHost` type.
+- A new **`devframe/internal`** entry (an explicitly-unstable cross-package
+  surface) holds the low-level primitives first-party integrations need:
+  `createContextRpcServer`, `DevframeAgentHost`, `coerceAgentPositionalArgs`,
+  the instance registry (`registerDevframeInstance` / `listLiveDevframeInstances`
+  + types), and `isObject` / `normalizeHttpServerUrl`. The hub, the inspect
+  plugin, and the host examples import from there. The remaining node internals
+  (diagnostics/services/views hosts, the streaming/shared-state/scope/settings
+  factories, `toDialableHost`/`formatHostForUrl`) stay fully internal (relative
+  imports only).
 
 ### ⚠️ One item to confirm during review
